# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "PVT OpenApi",
        "description": "L5 Swagger OpenApi description",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/temporary/upload_copy_person_senasir": {
            "post": {
                "tags": [
                    "IMPORTACION-IDS-PERSONAS-SENASIR"
                ],
                "summary": "PASO 1 COPIADO DE DATOS IMPORTACIÓN DE IDS SENASIR",
                "description": "Copiado de datos del archivo de afiliados senasir a la tabla copy_person_senasirs",
                "operationId": "upload_copy_person_senasir",
                "requestBody": {
                    "description": "Provide auth credentials",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "file": {
                                        "description": "file required",
                                        "type": "file",
                                        "example": "file"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/temporary/update_affiliate_id_senasir": {
            "post": {
                "tags": [
                    "IMPORTACION-IDS-PERSONAS-SENASIR"
                ],
                "summary": "PASO 2 ACTUALIZACIÓN DE IDS VERIFICACIÓN BAJO 5 CRITERIOS BASE",
                "description": "Actualización  de ids senasir a la tabla affiliates bajo 5 criterios automaticos",
                "operationId": "update_affiliate_id_senasir",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/temporary/update_affiliate_id_senasir_registration_and_identity_card": {
            "post": {
                "tags": [
                    "IMPORTACION-IDS-PERSONAS-SENASIR"
                ],
                "summary": "PASO 3 ACTUALIZACIÓN DE CRITERIOS 6 y 7 REVISION MANUAL",
                "description": "Actualización  de ids senasir a la tabla affiliates bajo 2 criterios de revisión manual",
                "operationId": "update_affiliate_id_senasir_registration_and_identity_card",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/temporary/create_affiliate_spouse_senasir": {
            "post": {
                "tags": [
                    "IMPORTACION-IDS-PERSONAS-SENASIR"
                ],
                "summary": "PASO 4 CREACIÓN DE AFILIADOS Y ESPOSAS SENASIR",
                "description": "Creación de afiliados y esposas senasir no encontrados en la tabla affiliates",
                "operationId": "create_affiliate_spouse_senasir",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/temporary/update_affiliate_data": {
            "post": {
                "tags": [
                    "IMPORTACION-IDS-PERSONAS-SENASIR"
                ],
                "summary": "PASO 5 ACTUALIZACIÓN DE DATOS DEL AFILIADO",
                "description": "Actualización de datos del afiliado, matricula, fecha de nacimiento, fecha de fallecimiento, ente gestor, y el estado del afiliado",
                "operationId": "update_affiliate_data",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}